Take 4

+++ Aufgabe 103 - Der Parkplatz/Interfaces/IFahrzeugZusatz.cs
@@ -8,6 +8,6 @@ namespace Aufgabe_103___Der_Parkplatz
 {
     public interface IFahrzeugZusatz
     {
-        int Zuladung { get; set; }
+        double Zuladung { get; set; }
     }
 }
